requirements.txt
Django==2.2.7
django-filter==2.2.0
django-rest-multiple-models==2.1.3
djangorestframework==3.10.3
mysql==0.0.2
mysqlclient==1.4.6
pytz==2019.3
sqlparse==0.3.0
